refactor: move `Destination` schemas to separate file
