+++ src/functions/ftp/external-poller/pollers/ftpPoller.ts
@@ -7,7 +7,7 @@ import { PutObjectCommand } from "@stedi/sdk-client-buckets";
 
 import { bucketClient } from "../../../../lib/buckets.js";
 import { FileDetails, ProcessingError, RemoteFileDetails } from "../types.js";
-import { DestinationBucket } from "../../../../lib/types/PartnerRouting.js";
+import { DestinationBucket } from "../../../../lib/types/Destination.js";
 import { ConnectionDetails } from "../../../../lib/types/RemoteConnectionConfig.js";
 import { RemotePoller } from "./remotePoller.js";
