+++ src/functions/ftp/external-poller/pollers/remotePoller.ts
@@ -1,5 +1,5 @@
 import { FileDetails, RemoteFileDetails } from "../types.js";
-import { DestinationBucket } from "../../../../lib/types/PartnerRouting.js";
+import { DestinationBucket } from "../../../../lib/types/Destination.js";
 import { ConnectionDetails } from "../../../../lib/types/RemoteConnectionConfig.js";
 
 export abstract class RemotePoller {
